Changes to includes and reflection API
