+++ Include/Containers/Map.h
@@ -9,11 +9,14 @@
 #include "Misc/Hash.h"
 #include "Misc/Utility.h"
 #include "Platform/Platform.h"
+#include "Reflection/TypeName.h"
+#include "Strings/StringView.h"
 
 #include <tsl/sparse_map.h>
 
 #include <cassert>
 #include <type_traits>
+
 
 
 namespace Rift
@@ -31,6 +34,7 @@ namespace Rift
 
 		using KeyType     = Key;
 		using ValueType   = Value;
+		using AllocatorType = Allocator;
 		using HashMapType = tsl::sparse_map<KeyType, ValueType, Hash<KeyType>,
 		    std::equal_to<KeyType>, STLAllocator<std::pair<Key, Value>, Allocator>>;
 
@@ -269,4 +273,59 @@ namespace Rift
 			map = Move(other.map);
 		}
 	};
+
+
+	template <typename T>
+	struct HasKeyType
+	{
+	private:
+		template <typename V>
+		static void Impl(decltype(typename V::KeyType(), int()));
+		template <typename V>
+		static bool Impl(char);
+
+	public:
+		static const bool value = std::is_void<decltype(Impl<T>(0))>::value;
+	};
+
+	template <typename T>
+	struct HasValueType
+	{
+	private:
+		template <typename V>
+		static void Impl(decltype(typename V::ValueType(), int()));
+		template <typename V>
+		static bool Impl(char);
+
+	public:
+		static const bool value = std::is_void<decltype(Impl<T>(0))>::value;
+	};
+
+
+	template <typename T>
+	inline constexpr bool IsMap()
+	{
+		// Check if we are dealing with a TAssetPtr
+		if constexpr (HasKeyType<T>::value && HasValueType<T>::value)
+		{
+			return IsSame<
+			    TMap<typename T::KeyType, typename T::ValueType, typename T::AllocatorType>, T>;
+		}
+		return false;
+	}
+
+	template <typename T>
+	constexpr StringView GetFullTypeName() requires(IsMap<T>())
+	{
+		constexpr auto preffix        = FixedString("TMap<");
+		constexpr auto suffix         = FixedString(">");
+		constexpr StringView itemName = GetTypeName<typename T::ItemType>();
+		return preffix + *itemName.data() + suffix;
+	}
+
+	template <typename T>
+	constexpr StringView GetTypeName() requires(IsMap<T>())
+	{
+		return "TMap";
+	}
 }    // namespace Rift
