+++ Include/CoreTypes.h
@@ -2,7 +2,7 @@
 #pragma once
 
 #include "Platform/Platform.h"
-#include "Reflection/Static/NativeType.h"
+#include "Reflection/Registry/NativeTypeBuilder.h"
 #include "Strings/Name.h"
 #include "Strings/String.h"
