+++ Include/Reflection/Reflection.h
@@ -3,6 +3,18 @@
 
 #include "PCH.h"
 
-#include "Reflection/Helpers.h"
-#include "Reflection/ReflectionMacros.h"
-#include "Reflection/TypeId.h"
+#include "Reflection/GetType.h"
+#include "Reflection/HasType.h"
+#include "Reflection/ReflectionTags.h"
+#include "Reflection/Registry/DataTypeBuilder.h"
+#include "Reflection/Registry/EnumTypeBuilder.h"
+#include "Reflection/Registry/NativeTypeBuilder.h"
+
+
+
+/**
+ * There are 2 types of reflected data structures:
+ *
+ * - STRUCT: Inheritance. Not virtual. Inlined in memory
+ * - CLASS: Inheritance. Virtual. Smart creation and destruction. Instanced in memory.
+ */
