+++ Include/Reflection/ReflectionTraits.h
@@ -1,13 +1,10 @@
 // Copyright 2015-2021 Piperift - All rights reserved
 #pragma once
 
-#include "PCH.h"
-
+#include "Containers/Array.h"
 #include "Object/BaseObject.h"
 #include "Reflection/Registry/TypeBuilder.h"
 #include "TypeTraits.h"
-
-#include <type_traits>
 
 
 
@@ -16,7 +13,10 @@ namespace Rift
 	struct Struct;
 
 	template <typename T>
-	class TAssetPtr;
+	inline constexpr bool IsEnum()
+	{
+		return std::is_enum<T>::value;
+	}
 
 	template <typename T>
 	inline constexpr bool IsStruct()
@@ -31,56 +31,13 @@ namespace Rift
 	}
 
 	template <typename T>
-	inline constexpr bool IsEnum()
+	inline constexpr bool IsArray()
 	{
-		return std::is_enum<T>::value;
-	}
-
-	template <typename T>
-	inline constexpr bool IsReflected()
-	{
-		if constexpr (IsArray<T>())
+		// Check if we are dealing with a TArray
+		if constexpr (HasItemType<T>::value)
 		{
-			return IsReflected<typename T::ItemType>();
+			return IsSame<TArray<typename T::ItemType>, T>;
 		}
-		return IsStruct<T>() || IsClass<T>() || IsEnum<T>();
-	}
-
-
-	template <typename T>
-	struct HasTypeBuilderDef
-	{
-	private:
-		template <typename V>
-		static void Impl(decltype(typename V::TypeBuilder(), int()));
-		template <typename V>
-		static bool Impl(char);
-
-	public:
-		static constexpr bool value = std::is_void<decltype(Impl<T>(0))>::value;
-	};
-
-	template <typename T>
-	constexpr bool HasTypeBuilder()
-	{
-		return HasTypeBuilderDef<T>::value && Derived<T::TypeBuilder, BaseBuilder, true>
-	}
-
-	template <typename T>
-	constexpr bool HasType() requires(IsStruct<T>() || IsClass<T>())
-	{
-		return HasTypeBuilder()
-	}
-
-	template <typename T>
-	constexpr bool HasType() requires(IsEnum<T>())
-	{
-		return Refl::TStaticEnumInitializer<T>::reflected;
-	}
-
-	template <typename T>
-	constexpr bool HasType()
-	{
-		return HasTypeBuilder();
+		return false;
 	}
 }    // namespace Rift
