+++ Include/Reflection/Static/DataType.h
@@ -6,7 +6,6 @@
 #include "Containers/Array.h"
 #include "Containers/Map.h"
 #include "CoreTypes.h"
-#include "Reflection/Helpers.h"
 #include "Reflection/ReflectionTags.h"
 #include "Reflection/Static/Property.h"
 #include "Reflection/Static/Type.h"
