+++ Src/Reflection/Registry/Registry.cpp
@@ -1,0 +1,13 @@
+// Copyright 2015-2021 Piperift - All rights reserved
+
+#include "Reflection/Registry/Registry.h"
+
+
+namespace Rift::Refl
+{
+	ReflectionRegistry& ReflectionRegistry::Get()
+	{
+		static ReflectionRegistry instance{};
+		return instance;
+	}
+}    // namespace Rift::Refl
