+++ Src/Reflection/Registry/TypeBuilder.cpp
@@ -1,16 +1,14 @@
 // Copyright 2015-2021 Piperift - All rights reserved
 
-#include "Reflection/Static/ReflectionRegistry.h"
+#include "Reflection/Registry/TypeBuilder.h"
+
+#include "Reflection/Registry/Registry.h"
+#include "Reflection/TypeId.h"
 
 
 namespace Rift::Refl
 {
-	ReflectionRegistry& ReflectionRegistry::Get()
-	{
-		static ReflectionRegistry instance{};
-		return instance;
-	}
-
+	TypeBuilder::TypeBuilder(TypeId id, Name name) : id{id}, name{name} {}
 
 	void TypeBuilder::Initialize()
 	{
@@ -26,5 +24,14 @@ namespace Rift::Refl
 		{
 			initializedType = Build();
 		}
+	}
+
+	TypeId TypeBuilder::GetId() const
+	{
+		return id;
+	}
+	Name TypeBuilder::GetName() const
+	{
+		return name;
 	}
 }    // namespace Rift::Refl
