+++ Src/Reflection/Static/Type.cpp
@@ -1,0 +1,3 @@
+// Copyright 2015-2021 Piperift - All rights reserved
+
+#include "Reflection/Static/Type.h"
